Fix reply defer not working
